Remove ExtendedMulticall
