--- contracts/utils/interfaces/IExtendedMulticall.sol
@@ -1,20 +1,0 @@
-// SPDX-License-Identifier: MIT
-pragma solidity ^0.8.0;
-
-import "./IMulticall.sol";
-
-interface IExtendedMulticall is IMulticall {
-    function getChainId() external view returns (uint256);
-
-    function getBalance(address account) external view returns (uint256);
-
-    function getBlockNumber() external view returns (uint256);
-
-    function getBlockTimestamp() external view returns (uint256);
-
-    function getBlockBasefee() external view returns (uint256);
-
-    function tryMulticall(bytes[] calldata data)
-        external
-        returns (bool[] memory success, bytes[] memory returndata);
-}
